pull out sub management into component
