+++ packages/dashboard/src/components/subscriptionManagement.tsx
@@ -1,0 +1,49 @@
+import { Checkbox, FormControlLabel, FormGroup } from "@mui/material";
+import { UserSubscriptionResource } from "isomorphic-lib/src/types";
+import React from "react";
+
+export interface SubscriptionManagementProps {
+  subscriptions: UserSubscriptionResource[];
+}
+
+export function SubscriptionManagement({
+  subscriptions,
+}: SubscriptionManagementProps) {
+  const initialSubscriptionManagementState = React.useMemo(
+    () =>
+      subscriptions.reduce<Record<string, boolean>>((acc, subscription) => {
+        acc[subscription.id] = true;
+        return acc;
+      }, {}),
+    [subscriptions]
+  );
+  const [state, setState] = React.useState(initialSubscriptionManagementState);
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setState({
+      ...state,
+      [event.target.name]: event.target.checked,
+    });
+  };
+
+  return (
+    <>
+      subscription management
+      <FormGroup>
+        {subscriptions.map((subscription) => (
+          <FormControlLabel
+            key={subscription.id}
+            control={
+              <Checkbox
+                checked={state[subscription.id] === true}
+                onChange={handleChange}
+                name={subscription.id}
+              />
+            }
+            label={subscription.name}
+          />
+        ))}
+      </FormGroup>
+    </>
+  );
+}
