+++ packages/dashboard/src/pages/subscription-management.page.tsx
@@ -1,4 +1,3 @@
-import { Checkbox, FormControlLabel, FormGroup } from "@mui/material";
 import {
   buildSubscriptionChangeEvent,
   generateSubscriptionHash,
@@ -9,19 +8,13 @@ import { SubscriptionChange } from "backend-lib/src/types";
 import { insertUserEvents } from "backend-lib/src/userEvents";
 import { UNAUTHORIZED_PAGE } from "isomorphic-lib/src/constants";
 import { schemaValidate } from "isomorphic-lib/src/resultHandling/schemaValidation";
-import {
-  SubscriptionParams,
-  UserSubscriptionResource,
-} from "isomorphic-lib/src/types";
+import { SubscriptionParams } from "isomorphic-lib/src/types";
 import { GetServerSideProps, NextPage } from "next";
 import Head from "next/head";
 import React from "react";
 
+import { SubscriptionManagementProps } from "../components/subscriptionManagement";
 import prisma from "../lib/prisma";
-
-interface SubscriptionManagementProps {
-  subscriptions: UserSubscriptionResource[];
-}
 
 export const getServerSideProps: GetServerSideProps<
   SubscriptionManagementProps
@@ -118,25 +111,6 @@ export const getServerSideProps: GetServerSideProps<
 
 const SubscriptionManagement: NextPage<SubscriptionManagementProps> =
   function SubscriptionManagement({ subscriptions }) {
-    const initialSubscriptionManagementState = React.useMemo(
-      () =>
-        subscriptions.reduce<Record<string, boolean>>((acc, subscription) => {
-          acc[subscription.id] = true;
-          return acc;
-        }, {}),
-      [subscriptions]
-    );
-    const [state, setState] = React.useState(
-      initialSubscriptionManagementState
-    );
-
-    const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      setState({
-        ...state,
-        [event.target.name]: event.target.checked,
-      });
-    };
-
     return (
       <>
         <Head>
@@ -144,22 +118,7 @@ const SubscriptionManagement: NextPage<SubscriptionManagementProps> =
           <meta name="description" content="Open Source Customer Engagement" />
         </Head>
         <main>
-          subscription management
-          <FormGroup>
-            {subscriptions.map((subscription) => (
-              <FormControlLabel
-                key={subscription.id}
-                control={
-                  <Checkbox
-                    checked={state[subscription.id] === true}
-                    onChange={handleChange}
-                    name={subscription.id}
-                  />
-                }
-                label={subscription.name}
-              />
-            ))}
-          </FormGroup>
+          <SubscriptionManagement subscriptions={subscriptions} />
         </main>
       </>
     );
